+++ packages/client/test/unit/Publisher.test.ts
@@ -16,7 +16,7 @@ describe('Publisher', () => {
             authentication,
             streamRegistry as any,
             undefined as any,
-            undefined as any
+            undefined as any,
         )
         const streamId = await streamIdBuilder.toStreamID('/test')
         await expect(async () => {